数据统计

+++ tlias-web-management/src/main/java/com/jy/service/impl/DeptServiceImpl.java
@@ -1,9 +1,13 @@
 package com.jy.service.impl;
 import com.itheima.pojo.Dept;
 import com.jy.mapper.DeptMapper;
+import com.jy.mapper.EmpMapper;
+import com.jy.pojo.DeptHasEmployeesException;
+import com.jy.pojo.Emp;
 import com.jy.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -12,13 +16,19 @@ import java.util.List;
 public class DeptServiceImpl implements DeptService {
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
     @Override
     public List<Dept> findAll() {
         return deptMapper.findAll();
     }
-
+    @Transactional(rollbackFor = {Exception.class})
     @Override
     public void deleteById(Integer id) {
+        List<Emp> emps = empMapper.getBydeptId(id);
+        if (!emps.isEmpty()) {
+            throw new DeptHasEmployeesException("对不起, 该部门下有员工, 不能直接删除");
+        }
         deptMapper.deleteById(id);
     }
     @Override
